requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic[email]==2.5.3

# Azure Services
azure-storage-blob==12.19.0
azure-ai-formrecognizer==3.3.2
azure-search-documents==11.4.0
azure-identity==1.15.0
openai==1.10.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1

# Background tasks (optional, for async processing)
celery==5.3.6
redis==5.0.1

# Production server
gunicorn==21.2.0
pydantic-settings==2.1.0
